+++ src/main/java/me/axlerogue/horrifyinglanterns/api/ability/AbilityType.java
@@ -2,5 +2,7 @@ package me.axlerogue.horrifyinglanterns.api.ability;
 
 public enum AbilityType {
     BURST,
-    LEECH
+    LEECH,
+    SUMMON,
+    WRATH
 }