+++ src/main/java/me/axlerogue/horrifyinglanterns/api/client/renderer/BaseLivingEntityRenderer.java
@@ -1,0 +1,15 @@
+package me.axlerogue.horrifyinglanterns.api.client.renderer;
+
+import net.minecraft.client.model.EntityModel;
+import net.minecraft.client.renderer.entity.EntityRendererProvider;
+import net.minecraft.client.renderer.entity.LivingEntityRenderer;
+import net.minecraft.world.entity.LivingEntity;
+
+/**
+ * Base class for all horrifying lantern living entity renderers.
+ */
+public abstract class BaseLivingEntityRenderer<T extends LivingEntity, M extends EntityModel<T>> extends LivingEntityRenderer<T, M> {
+    protected BaseLivingEntityRenderer(EntityRendererProvider.Context context, M model, float shadowRadius) {
+        super(context, model, shadowRadius);
+    }
+}